ComicsLibrary/requirements.txt
fastapi
motor
uvicorn
python-dotenv
astroid